refactor(eslint): move project type detection to a separate utility function
